user model and roles added
