--- ComicHub/src/main/java/org/comicshub/comichub/Models/User.java
@@ -1,23 +1,0 @@
-package org.comicshub.comichub.Models;
-
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.awt.*;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class User {
-
-    private long id;
-    private String username;
-    private String password;
-    private String email;
-    private Image avatar;
-
-}
