+++ ComicHub/src/main/java/org/comicshub/comichub/Security/Role.java
@@ -1,0 +1,7 @@
+package org.comicshub.comichub.Security;
+
+public enum Role {
+
+    ROLE_GUEST, ROLE_USER, ROLE_ADMIN
+
+}
